put face lock at service level
