+++ src/faceAgent.cpp
@@ -113,8 +113,6 @@ void FaceAgent::getUserFaces(const std::string &appName,
     const std::string &groupName,
     const std::string &userName,
     std::map<std::string, std::shared_ptr<ImageFace>> &faceMap) {
-  RLockMethod md;
-  RWLockGuard guard(md, &lock_);
   auto app = getAppFace(appName);
   if (app == nullptr) {
     return;
@@ -130,8 +128,6 @@ void FaceAgent::getUserFaces(const std::string &appName,
 
 int FaceAgent::delPerson(const PersonFace &face) {
   std::string appName = (face.appName == "" ? DEFAULT_APP_NAME : face.appName);
-  WLockMethod md;
-  RWLockGuard guard(md, &lock_);
   auto app = getAppFace(appName);
   if (app == nullptr) {
     return -1;
@@ -147,8 +143,6 @@ int FaceAgent::delPerson(const PersonFace &face) {
 
 int FaceAgent::delPersonFace(const PersonFace &face) {
   std::string appName = (face.appName == "" ? DEFAULT_APP_NAME : face.appName);
-  WLockMethod md;
-  RWLockGuard guard(md, &lock_);
   auto app = getAppFace(appName);
   if (app == nullptr) {
     return -1;
@@ -168,8 +162,6 @@ int FaceAgent::delPersonFace(const PersonFace &face) {
 
 int FaceAgent::addPersonFace(const PersonFace &face) {
   std::string appName = (face.appName == "" ? DEFAULT_APP_NAME : face.appName);
-  WLockMethod md;
-  RWLockGuard guard(md, &lock_);
   auto app = getAppFace(appName);
   if (app == nullptr) {
     addAppFace(appName);
@@ -201,8 +193,6 @@ int FaceAgent::addPersonFace(const PersonFace &face) {
 }
 
 void FaceAgent::getDefaultPersonFaces(std::list<PersonFace> &faces) {
-  WLockMethod md;
-  RWLockGuard guard(md, &lock_);
   auto app = getAppFace(DEFAULT_APP_NAME);
   if (app == nullptr) {
     return;
